% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{bmr.schofield}
\alias{bmr.schofield}
\title{Basal Metabolic Rate as per Schofield's equations}
\usage{
bmr.schofield(age, sex = c("male", "female"), weight)
}
\arguments{
\item{age}{in years}

\item{sex}{"male" or "female"}

\item{weight}{in kg}
}
\value{
BMR, in kcal/day
}
\description{
Basal Metabolic Rate as per Schofield's equations
}
