% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{clear.labels}
\alias{clear.labels}
\title{supprime tous les labels de HMisc pour eviter les problemes avec dplyr}
\usage{
clear.labels(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
supprime tous les labels de HMisc pour eviter les problemes avec dplyr
}
