% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{cox.plot.zph}
\alias{cox.plot.zph}
\title{cox.plot.zph
A wrapper around `survminer::ggcoxzph`}
\usage{
cox.plot.zph(cox.fit, zph.transform, vars, vars_pval = "all",
  size = 1.25)
}
\arguments{
\item{cox.fit}{a `coxph` object}

\item{zph.transform}{the `transform` parameter to pass to `survival::cox.zph`}

\item{vars}{a character vector of model variables, either extact or not}

\item{vars_pval}{a numeric value to draw only if zph test pvalue is under. Override `vars` if used.}

\item{size}{the size of the beta's blue line.}
}
\value{
the ggplot
}
\description{
cox.plot.zph
A wrapper around `survminer::ggcoxzph`
}
\examples{
library(survival)
library(survminer)

fit <- coxph(Surv(time, event) ~ molecular_group + CCND1 +
                                 CRIM1 + DEPDC1 + IRF4 + TP53 +
                                 WHSC1, data=myeloma)
cox.plot.zph(fit, "km")
cox.plot.zph(fit, "km", vars="molecular_group")
cox.plot.zph(fit, "km", vars=c("CCND1", "molecular_groupCyclin D-2",
                               "molecular_groupHyperdiploid", "molecular_groupProliferation"))
cox.plot.zph(fit, "km", vars_pval=0.2)

###To plot the result
gp=gpar(fontsize=20,font=3)
f = function(p) paste0("Global Schoenfeld Test p:", round(attr(p, "global_pval"),3))
p1 = cox.plot.zph(fit, "km")
p1 = arrangeGrob(grobs = p1,
                 top = textGrob(f(p1),gp=gp))
ggsave(filename, p1, scale = 3)


}
