% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{dan.na.table}
\alias{dan.na.table}
\title{Intersections des NA
donne les intersection des NA
utiliser apres un select sur une dataframe
Voir aussi mice::md.pattern}
\usage{
dan.na.table(df.na, pnames = FALSE)
}
\description{
Intersections des NA
donne les intersection des NA
utiliser apres un select sur une dataframe
Voir aussi mice::md.pattern
}
