% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{dan.operators}
\alias{dan.operators}
\alias{\%==\%}
\alias{\%!=\%}
\title{Operators not indluded in R}
\usage{
a \%==\% b

a \%!=\% b
}
\arguments{
\item{x}{=X}

\item{y}{=Y}
}
\value{
na-proof equality or difference

never NA
}
\description{
Exactly equals

Exactly inequal
}
\examples{
data.frame(a=c(NA,1,NA,1,1), b=c(NA,NA,1,1,2)) \%>\% mutate(
equal = a==b,
different = a!=b,
real_equal = a\%==\%b,
real_different = a\%!=\%b
)
}
