% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{factor_by_freq}
\alias{factor_by_freq}
\title{Cast to factor, ordering by frequency}
\usage{
factor_by_freq(x)
}
\arguments{
\item{x}{a character or a factor}
}
\value{
a factor, ordered by frquency
}
\description{
Cast to factor, ordering by frequency
}
\examples{
x=iris$Sepal.Length \%>\% cut(c(-Inf, 4.5, 5.4, Inf))
x \%>\% table
x \%>\% factor_by_freq \%>\% table
x \%>\% factor_by_freq \%>\% levels
}
