% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{formula_to_varlist}
\alias{formula_to_varlist}
\title{Formula to variable list}
\usage{
formula_to_varlist(fo)
}
\arguments{
\item{fo}{a single formula or a string to be parsed as formula.}
}
\value{
a character vector of formula members, with used functions as attribute
}
\description{
Formula to variable list
}
\examples{
fo = formula("Surv(start, stop, event) ~ X + log(Y) + Z + rcs(W, 3)")
formula_to_varlist(fo)
}
