% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{geom_point_and_error}
\alias{geom_point_and_error}
\title{à ajouter à un ggplot
#TODO ajouter sd}
\usage{
geom_point_and_error(type = c("sem", "ic"))
}
\arguments{
\item{type}{sem ou ic}
}
\description{
à ajouter à un ggplot
#TODO ajouter sd
}
