% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get.formula}
\alias{get.formula}
\title{get.formula}
\usage{
get.formula(LHS, RHS, ...)
}
\description{
get.formula
}
\examples{
get.formula(c("score.pnns4.C"), c("age_inclus_dan", "sexe.cl"))
get.formula("score.pnns4.C", "age_inclus_dan")
}
