% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{getQuantiles}
\alias{getQuantiles}
\title{Recode une variable en quantiles, en groupant ou non}
\usage{
getQuantiles(x, data, by = FALSE, n = 5, letter = "Q")
}
\arguments{
\item{x}{la variable à recoder}

\item{data}{data.frame}

\item{by}{la variable de groupement (ou FALSE sinon)}
}
\description{
en quintiles non groupes par défaut
}
\examples{
getQuantiles(Sepal.Length, by=Species, data=iris)
getQuantiles(Sepal.Length, by=FALSE, data=iris)
getQuantiles(Sepal.Length, by=FALSE, n=3, letter="T", data=iris)
}
