% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ggcoxfunctional_bak}
\alias{ggcoxfunctional_bak}
\title{Amelioration of survminer::ggcoxfunctional
added 2 lines to remove all non-numeric variables and avoid errors}
\usage{
ggcoxfunctional_bak(formula, data = NULL, fit, iter = 0, f = 0.6,
  select = NULL, point.col = "red", point.size = 1,
  point.shape = 19, point.alpha = 1, xlim = NULL, ylim = NULL,
  ylab = "Martingale Residuals \\nof Null Cox Model", title = NULL,
  caption = NULL, ggtheme = theme_survminer(), ...)
}
\description{
This won't work with \code{rcs}, but it's allright since it's meant to correct for linearity.
}
\details{
To save the result : \code{
p = arrangeGrob(grobs = ggcoxfunctional(fit=m1, data=data),
                left = textGrob("Martingale Residuals of Null Cox Model",gp=gp))
ggsave(filepath, p)
}
}
\seealso{
\code{\link[survminer]{ggcoxfunctional}} for actual help.
}
