% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{percentiles}
\alias{percentiles}
\title{Recode une variable en percentiles}
\usage{
percentiles(x, n, letter = "Q")
}
\arguments{
\item{x}{un vecteur à transformer}

\item{n}{le nombre de groupes (5 pour des quintiles)}

\item{letter}{le prefixe des groupes}
}
\description{
Recode une variable en percentiles
}
\examples{
percentiles(iris$Sepal.Length, 3, "T")
percentiles(iris$Sepal.Length, 5, "Q")
}
