% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{+}
\alias{+}
\title{Addition2
Reecriture de l'addition pour prendre en compte les characters (paste0)}
\usage{
"+"(x, y)
}
\description{
Addition2
Reecriture de l'addition pour prendre en compte les characters (paste0)
}
\examples{
"a"+"b"
}
\seealso{
\url{https://stackoverflow.com/a/18175871/3888000}
}
