% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{pval.star}
\alias{pval.star}
\title{affiche les etoiles en fonction de la significativite de la pvalue}
\usage{
pval.star(pval, digits = 3, limit = 1e-04, onlystar = F)
}
\description{
affiche les etoiles en fonction de la significativite de la pvalue
}
