% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{recursive_length}
\alias{recursive_length}
\title{Recursive length of a list of lists}
\usage{
recursive_length(x)
}
\arguments{
\item{x}{a list of lists}
}
\description{
count the number of terminal children (not list) of a list of lists
}
